requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
iso8601==0.1.14
m3u8==0.8.0
natsort==7.1.1
pycryptodome==3.10.1
requests==2.25.1
soupsieve==2.2.1
tqdm==4.59.0
urllib3==1.26.4
